requirements.txt
beautifulsoup4==4.12.3
cohere==5.3.2
config==0.5.1
easygui==0.98.3
Flask==2.1.0
langchain==0.1.16
langchain_community==0.0.34
nltk==3.8.1
pathvalidate==3.2.0
pypdf==4.2.0
PyPDF2==3.0.1
PyYAML==6.0
PyYAML==6.0.1
qdrant_client==1.9.0
Requests==2.31.0
scikit_learn==1.3.0
spacy==3.7.4
textacy==0.13.0
textdistance==4.6.1
xhtml2pdf==0.2.11
